LayoutTests/inspector/debugger/setShouldBlackboxURL-exact-case-sensitive-expected.txt

Tests Debugger.setShouldBlackboxURL.


== Running test suite: Debugger.setShouldBlackboxURL
-- Running test case: Debugger.setShouldBlackboxURL.ExactCaseSensitive.stepOver
Evaluating 'createScripts("ExactCaseSensitive_StepOver")'...
Blackboxing (case sensitive) 'exactcasesensitive_stepover_inner.js'...
Blackboxing (case sensitive) 'ExactCaseSensitive_StepOver_Middle.js'...
Setting breakpoint in 'ExactCaseSensitive_StepOver_Inner.js'...
Evaluating 'ExactCaseSensitive_StepOver_Outer(10)'...

PAUSED: 'Breakpoint' at 'ExactCaseSensitive_StepOver_Inner:3:1'.
{
  "breakpointId": "ExactCaseSensitive_StepOver_Inner.js:3:0"
}
Stepping over...

PAUSED: 'BlackboxedScript' at 'ExactCaseSensitive_StepOver_Outer:3:1'.
{
  "originalReason": "other",
  "originalData": {
    "breakpointId": "ExactCaseSensitive_StepOver_Inner.js:3:0"
  }
}
Stepping over...

Resuming...
PASS: Resumed.
PASS: Should pause in 'ExactCaseSensitive_StepOver_Inner'.
PASS: Should not pause in 'ExactCaseSensitive_StepOver_Middle'.


-- Running test case: Debugger.setShouldBlackboxURL.ExactCaseSensitive.PauseInCaller
Evaluating 'createScripts("ExactCaseSensitive_PauseInCaller")'...
Blackboxing (case sensitive) 'ExactCaseSensitive_PauseInCaller_Inner.js'...
Blackboxing (case sensitive) 'exactcasesensitive_pauseincaller_middle.js'...
Setting breakpoint in 'ExactCaseSensitive_PauseInCaller_Inner.js'...
Evaluating 'ExactCaseSensitive_PauseInCaller_Outer(10)'...

PAUSED: 'BlackboxedScript' at 'ExactCaseSensitive_PauseInCaller_Middle:3:1'.
{
  "originalReason": "Breakpoint",
  "originalData": {
    "breakpointId": "ExactCaseSensitive_PauseInCaller_Inner.js:2:0"
  }
}
Stepping over...

PAUSED: 'other' at 'ExactCaseSensitive_PauseInCaller_Outer:3:1'.
Stepping over...

Resuming...
PASS: Resumed.
PASS: Should not pause in 'ExactCaseSensitive_PauseInCaller_Inner'.
PASS: Should pause in 'ExactCaseSensitive_PauseInCaller_Middle'.


-- Running test case: Debugger.setShouldBlackboxURL.ExactCaseSensitive.PauseInCallee
Evaluating 'createScripts("ExactCaseSensitive_PauseInCallee")'...
Blackboxing (case sensitive) 'exactcasesensitive_pauseincallee_middle.js'...
Blackboxing (case sensitive) 'ExactCaseSensitive_PauseInCallee_Outer.js'...
Setting breakpoint in 'ExactCaseSensitive_PauseInCallee_Outer.js'...
Evaluating 'ExactCaseSensitive_PauseInCallee_Outer(10)'...

PAUSED: 'BlackboxedScript' at 'ExactCaseSensitive_PauseInCallee_Middle:2:4'.
{
  "originalReason": "Breakpoint",
  "originalData": {
    "breakpointId": "ExactCaseSensitive_PauseInCallee_Outer.js:2:0"
  }
}
Stepping over...

PAUSED: 'other' at 'ExactCaseSensitive_PauseInCallee_Middle:3:1'.
Stepping over...

Resuming...
PASS: Resumed.
PASS: Should pause in 'ExactCaseSensitive_PauseInCallee_Middle'.
PASS: Should not pause in 'ExactCaseSensitive_PauseInCallee_Outer'.


-- Running test case: Debugger.setShouldBlackboxURL.ExactCaseSensitive.SourceRange
Evaluating 'createScripts("ExactCaseSensitive_Range")'...
Blackboxing (case sensitive) (regex) 'exactcasesensitive_range_multiple.js'...
Blackboxing (case sensitive) 'ExactCaseSensitive_Range_Multiple.js' 3:0-4:0...
Blackboxing (case sensitive) 'ExactCaseSensitive_Range_Multiple.js' 4:5-4:18...
Blackboxing (case sensitive) 'ExactCaseSensitive_Range_Multiple.js' 5:1-5:2...
Setting breakpoint in 'ExactCaseSensitive_Range_Multiple.js'...
Evaluating 'ExactCaseSensitive_Range_Multiple(10)'...

PAUSED: 'Breakpoint' at 'ExactCaseSensitive_Range_Multiple:2:4'.
{
  "breakpointId": "ExactCaseSensitive_Range_Multiple.js:2:0"
}
Stepping over...

PAUSED: 'BlackboxedScript' at 'ExactCaseSensitive_Range_Multiple:4:4'.
{
  "originalReason": "other",
  "originalData": {
    "breakpointId": "ExactCaseSensitive_Range_Multiple.js:2:0"
  }
}
Stepping over...

Resuming...
PASS: Resumed.
PASS: Should pause in 'ExactCaseSensitive_Range_Multiple'.
